+++ python/src/ouster/pcap/__init__.py
@@ -2,7 +2,8 @@
 # flake8: noqa: F401 (unused imports)
 
 from .pcap import Pcap
-from .pcap import _replay
-from .pcap import _guess_ports
 from .pcap import record
-from .pcap import _pcap_info
+from .pcap import _guess_ports
+from .pcap import _packet_info_stream
+from .pcap import _replay
+from .pcap import _stream_info