+++ python/tests/test_pcap.py
@@ -1,6 +1,7 @@
+from collections import defaultdict
 import os
 from random import getrandbits, shuffle, random
-from typing import Iterator
+from typing import Dict, Iterable, Iterator, List
 from itertools import chain, islice
 
 from more_itertools import roundrobin
@@ -68,10 +69,10 @@ def use_sll() -> int:
 
 
 @pytest.fixture
-def pcap_path(meta, n_packets, use_sll, tmpdir):
+def pcap_path(meta, n_packets, use_sll, tmpdir) -> Iterable[str]:
     file_path = os.path.join(tmpdir, "pcap_test.pcap")
 
-    packets = islice(
+    packets: Iterable[client.Packet] = islice(
         roundrobin(random_lidar_packets(meta), random_imu_packets(meta)),
         n_packets)
 
@@ -81,7 +82,7 @@ def pcap_path(meta, n_packets, use_sll, tmpdir):
 
 
 @pytest.fixture
-def pcap_obj(meta, pcap_path):
+def pcap_obj(meta, pcap_path) -> Iterable[pcap.Pcap]:
     pc = pcap.Pcap(pcap_path, meta)
     yield pc
     pc.close()
@@ -89,8 +90,10 @@ def pcap_obj(meta, pcap_path):
 
 @pytest.mark.parametrize('test_key', ['legacy-2.0'])
 @pytest.mark.parametrize('n_packets', [0])
-def test_pcap_read_empty(pcap_obj) -> None:
-    """Check that reading an empty pcap yields an empty list."""
+def test_pcap_read_empty(meta, pcap_path) -> None:
+    """Check reading with fully specified ports doesn't fail."""
+    pcap_obj = pcap.Pcap(pcap_path, meta)
+    assert pcap_obj.ports == (0, 0)
     assert list(pcap_obj) == []
 
 
@@ -98,41 +101,26 @@ def test_pcap_read_empty(pcap_obj) -> None:
 @pytest.mark.parametrize('n_packets', [10])
 def test_pcap_read_10(pcap_obj) -> None:
     """Check that reading a test pcap produces the right number of packets."""
+    assert pcap_obj.ports == (7502, 7503)
     assert len(list(pcap_obj)) == 10
 
 
 @pytest.mark.parametrize('test_key', ['legacy-2.0'])
 @pytest.mark.parametrize('n_packets', [10])
 def test_pcap_info_10(pcap_path, meta) -> None:
-    """Check that reading a test pcap produces the right number of packets."""
-    res = pcap.Pcap(pcap_path, meta)
-    ports = {}
-    sizes = {}
-    encap = {}
-    net = {}
-    af = {}
-    for item in pcap._pcap_info(pcap_path):
-        if item.dst_port not in ports:
-            ports[item.dst_port] = 0
+    """Test reading packet headers with private helper."""
+    ports: Dict[int, int] = defaultdict(int)
+    sizes: Dict[int, int] = defaultdict(int)
+    encap: Dict[int, int] = defaultdict(int)
+    net: Dict[int, int] = defaultdict(int)
+    af: Dict[int, int] = defaultdict(int)
+
+    for item in pcap._packet_info_stream(pcap_path):
         ports[item.dst_port] += 1
-
-        if item.payload_size not in sizes:
-            sizes[item.payload_size] = 0
         sizes[item.payload_size] += 1
-
-        if item.encapsulation_protocol not in encap:
-            encap[item.encapsulation_protocol] = 0
         encap[item.encapsulation_protocol] += 1
-
-        if item.network_protocol not in net:
-            net[item.network_protocol] = 0
         net[item.network_protocol] += 1
-
-        if item.ip_version not in af:
-            af[item.ip_version] = 0
         af[item.ip_version] += 1
-    # default ports
-    assert res._lidar_port == 7502
 
     # roundrobin -> 5 of each
     assert ports == {7502: 5, 7503: 5}
@@ -146,14 +134,12 @@ def test_pcap_info_10(pcap_path, meta) -> None:
 @pytest.mark.parametrize('n_packets', [10])
 @pytest.mark.parametrize('use_sll', [True, False])
 def test_pcap_info_encap_proto(pcap_path, use_sll) -> None:
-    """Check that reading a test pcap produces the right number of packets."""
-    encap = {}
+    """Test reading/writing pcaps with different encapsulation."""
+    encap: Dict[int, int] = defaultdict(int)
 
-    for item in pcap._pcap_info(pcap_path):
-
-        if item.encapsulation_protocol not in encap:
-            encap[item.encapsulation_protocol] = 0
+    for item in pcap._packet_info_stream(pcap_path):
         encap[item.encapsulation_protocol] += 1
+
     proto = SLL_PROTO if use_sll else ETH_PROTO
     assert encap == {proto: 10}
 
@@ -187,11 +173,11 @@ def test_pcap_read_closed(pcap_obj) -> None:
     pytest.param(1, 1, id="one each ether"),
     pytest.param(20, 20, id="multi each ether"),
 ])
-def test_read_write_lidar_imu(n_lidar, n_imu, meta, tmpdir):
+def test_read_write_lidar_imu(n_lidar, n_imu, meta, tmpdir) -> None:
     """Test that random packets read back from pcap are identical."""
     lidar_packets = islice(random_lidar_packets(meta), n_lidar)
     imu_packets = islice(random_imu_packets(meta), n_imu)
-    in_packets = list(chain(lidar_packets, imu_packets))
+    in_packets: List[client.Packet] = list(chain(lidar_packets, imu_packets))
 
     shuffle(in_packets)
 
@@ -264,15 +250,11 @@ def test_no_timestamp_read_write(meta, tmpdir):
         pytest.param(0, 0, 10, 10, id="mixed: imu"),
         pytest.param(10, 0, 0, 10, id="mixed: lidar ts, imu no ts"),
         pytest.param(0, 10, 10, 10, id="mixed: lidar no ts, imu ts"),
-        pytest.param(10,
-                     10,
-                     10,
-                     10,
-                     id="mixed: lidar ts, lidar no ts, imu ts, imu no ts"),
+        pytest.param(10, 10, 10, 10, id="mixed: all"),
     ])
 def test_mixed_timestamp_write(n_lidar_timestamp, n_lidar_no_timestamp,
                                n_imu_timestamp, n_imu_no_timestamp, meta,
-                               tmpdir):
+                               tmpdir) -> None:
 
     lidar_timestamp_packets = islice(
         random_lidar_packets(meta, random_time=RANDOM_FLOAT),
@@ -285,7 +267,7 @@ def test_mixed_timestamp_write(n_lidar_timestamp, n_lidar_no_timestamp,
     imu_no_timestamp_packets = islice(
         random_imu_packets(meta, random_time=NO_RANDOM_TIME),
         n_imu_no_timestamp)
-    in_packets = list(
+    in_packets: List[client.Packet] = list(
         chain(lidar_timestamp_packets, lidar_no_timestamp_packets,
               imu_timestamp_packets, imu_no_timestamp_packets))
 
@@ -301,19 +283,18 @@ def test_mixed_timestamp_write(n_lidar_timestamp, n_lidar_no_timestamp,
         pcap.record(in_packets, file_path)
 
 
-@pytest.mark.parametrize('test_key', ['legacy-2.0'])
-def test_write_nonsensical_packet_type(meta, tmpdir):
+def test_write_nonsensical_packet_type(tmpdir) -> None:
     file_path = os.path.join(tmpdir, "pcap_test.pcap")
 
     in_packets = [42]
     with pytest.raises(ValueError):
-        pcap.record(in_packets, file_path)
+        pcap.record(in_packets, file_path)  # type: ignore
 
     assert not os.path.exists(file_path), "Didn't clean up empty file"
 
 
 @pytest.mark.parametrize('test_key', ['legacy-2.0'])
-def test_lidar_guess_error(meta, tmpdir):
+def test_lidar_guess_error(meta, tmpdir) -> None:
     packets = islice(random_lidar_packets(meta), 2)
     file_path = os.path.join(tmpdir, "pcap_test.pcap")
 
@@ -326,12 +307,13 @@ def test_lidar_guess_error(meta, tmpdir):
     finally:
         _pcap.record_uninitialize(handle)
 
-    with pytest.raises(ValueError):
-        pcap.Pcap(file_path, meta)
+    source = pcap.Pcap(file_path, meta)
+    assert len(source._guesses) > 1
+    assert len(list(source)) == 1
 
 
 @pytest.mark.parametrize('test_key', ['legacy-2.0'])
-def test_imu_guess_error(meta, tmpdir):
+def test_imu_guess_error(meta, tmpdir) -> None:
     packets = islice(random_imu_packets(meta), 2)
     file_path = os.path.join(tmpdir, "pcap_test.pcap")
 
@@ -344,5 +326,6 @@ def test_imu_guess_error(meta, tmpdir):
     finally:
         _pcap.record_uninitialize(handle)
 
-    with pytest.raises(ValueError):
-        pcap.Pcap(file_path, meta)
+    source = pcap.Pcap(file_path, meta)
+    assert len(source._guesses) > 1
+    assert len(list(source)) == 1